Move SASL mechanisms to separate files
